refactor(#762): consolidate glrender tests — eliminate 35 duplicates
Merge RenderContextOpacityStackTest + RenderContextExtendedTest +
RenderContextExternalServiceTest → RenderContextTest (59 unique tests,
down from 79 across 3 files).

Remove 6 duplicate scaled-count tests from PickContextConstantsTest
(already in ContextScaledCountTest).

Remove 9 duplicate PickResult tests from PickParametersDataTest
(isFrontFacing, subElement, position, null checks were tested twice).

All 899 tests pass. Net: -318 lines, -35 duplicate tests.
